perf(build): transpile internal packages
affects: @haaretz/haaretz.co.il

Apps now import the source of packages internal to the monorepo
and transpile them on build or when running the dev server.

This automatically enables better tree-shaking, and will enable future
optimizations through transpilation tragets etc. to be set in a single
place for each app, and on a per-app basis.

This should also greatly improve DX, as changes in the dependant
packages are watched when running the app's dev server, and there is no
need to build individual packages on every change. Live reload it then
also supported out of the box.

Naturally, changes to the server packages such as `htz-react-server` and
`graphql-server` cannot be hot-reloaded. They still require a build step
on every change and the app to be restarted.